requirements.txt
asgiref==3.5.2
black==22.10.0
click==8.1.3
colorama==0.4.6
Django==4.1.2
mypy-extensions==0.4.3
mysql-connector-python==8.0.31
pathspec==0.10.1
platformdirs==2.5.2
protobuf==3.20.1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.5
